backend/requirements.txt
Flask~=3.0.3
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
Faker~=25.8.0
PyJWT~=2.8.0